+++ llm_assistant/tests/__init__.py
@@ -2,3 +2,4 @@ from . import test_prompt_arguments
 from . import test_loop_control
 from . import test_pause
 from . import test_llm_retry
+from . import test_stream_trace_sink